+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## [unreleased]
+## unreleased
 
+- feature(Appender): migrated `Appender`, `CrashReportingAppender`, `DebugAppender` and `FileAppender` to Kotlin
+- feature(Formatter): migrated `Formatter`, `DefaultFormatter` and `SafeFormatter` to Kotlin. Warning signature of `Formatter.format` has changed for parameter `args`.
+- feature(Logger): migrated Logger to Kotlin
 - feature(annoation): migrated to androidx library
 - feature(gradle): migrate to version catalogs
 - fix(publication): cleaned up code added missing description and bound publication info to project instance
@@ -7,4 +10,4 @@
 - feature(JaCoCo): migrated to maintained jacoco dependency, removed sonar gradle plugin in favor of sonar scanner
 - feature(gradle): updated gradle wrapper and gradle scripts
 
-## [1.4] - 2019-11-21
+## 1.4 - 2019-11-21